ACS-7577 move edit and create folder directives to demo-shell [ci:force]
