+++ demo-shell/src/app/folder-directive/folder-create.directive.spec.ts
@@ -22,8 +22,8 @@ import { By } from '@angular/platform-browser';
 import { Subject, of } from 'rxjs';
 import { FolderCreateDirective } from './folder-create.directive';
 import { Node } from '@alfresco/js-api';
-import { ContentTestingModule } from '../testing/content.testing.module';
-import { ContentService } from '../common/services/content.service';
+import { ContentService } from '@alfresco/adf-content-services';
+import { ContentTestingModule } from 'lib/content-services/src/lib/testing/content.testing.module';
 
 @Component({
     template: ` <div [adf-create-folder]="parentNode" (success)="success($event)" title="create-title" [nodeType]="'cm:my-little-pony'"></div>`
