+++ demo-shell/src/app/folder-directive/folder-edit.directive.md
@@ -1,7 +1,7 @@
 ---
 Title: Folder Edit directive
 Added: v2.0.0
-Status: Active
+Status: Deprecated
 Last reviewed: 2019-01-16
 ---
