+++ demo-shell/src/app/folder-directive/index.ts
