+++ lib/content-services/src/lib/content.module.ts
@@ -33,7 +33,6 @@ import { ContentNodeSelectorModule } from './content-node-selector/content-node-
 import { ContentNodeShareModule } from './content-node-share/content-node-share.module';
 import { ContentDirectiveModule } from './directives/content-directive.module';
 import { DialogModule } from './dialogs/dialog.module';
-import { FolderDirectiveModule } from './folder-directive/folder-directive.module';
 import { ContentMetadataModule } from './content-metadata/content-metadata.module';
 import { PermissionManagerModule } from './permission-manager/permission-manager.module';
 import { TreeViewModule } from './tree-view/tree-view.module';
@@ -70,7 +69,6 @@ import { ContentAuthLoaderService } from './auth-loader/content-auth-loader.serv
         ContentNodeSelectorModule,
         ContentNodeShareModule,
         ContentMetadataModule,
-        FolderDirectiveModule,
         ContentDirectiveModule,
         PermissionManagerModule,
         VersionManagerModule,
@@ -98,7 +96,6 @@ import { ContentAuthLoaderService } from './auth-loader/content-auth-loader.serv
         ContentNodeShareModule,
         ContentMetadataModule,
         DialogModule,
-        FolderDirectiveModule,
         ContentDirectiveModule,
         PermissionManagerModule,
         VersionManagerModule,
